refactor: implement integer-only persistence for monetary values
- Updated SQL migrations (001, 003, 004, 006, 007, 010) to use BIGINT for monetary columns.
- Refactored src/funding/ (deposit, withdraw, service, handlers) to use i64 for internal logic and storage.
- Implemented scaling/unscaling logic at the API boundary using rust_decimal.
- Updated Admin Dashboard models (Python) to use BigInteger for consistency.
- Fixed mock deposit idempotency handling in handlers.rs (now returns 200 OK on duplicate).
- Updated .sqlx metadata for the new schema.
- Adjusted scripts and CI config for dev environment consistency.

+++ admin/models/tables.py
@@ -87,7 +87,7 @@ class VIPLevel(Base):
     
     level: Mapped[int] = mapped_column(SmallInteger, primary_key=True)
     discount_percent: Mapped[int] = mapped_column(SmallInteger, nullable=False, default=100)  # 100 = no discount
-    min_volume: Mapped[Optional[Decimal]] = mapped_column(Numeric(30, 8), default=0)
+    min_volume: Mapped[Optional[int]] = mapped_column(BigInteger, default=0)
     description: Mapped[Optional[str]] = mapped_column(String(64))
